authentication/requirements.txt
argon2-cffi>=25.1.0
python-dotenv>=0.9.9
fastapi>=0.121.1
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
pyjwt>=2.10.1
python-multipart>=0.0.20
typing-extensions>=4.15.0
uvicorn>=0.38.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
